pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "commitr"
version = "1.0.0"
description = "Git Commit Message Assistant"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE.TXT" }
authors = [
  { name = "Theodore Van Dusen", email = "theodore.vandusen@hotmail.com" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
]
dependencies = ["click", "requests", "openai", "python-dotenv"]

[project.scripts]
commitr = "pkg.main:main"

[tool.setuptools]
packages = ["pkg"]
